Split the nested components from this single file into individual component files.
