+++ src/apps/settings/src/app/views/Fonts/Browse/index.tsx
@@ -1,249 +1,23 @@
-import { useEffect, useState, FC, useMemo, useRef, useCallback } from "react";
-
+import { useEffect, useState, useMemo, useRef } from "react";
 import TextField from "@mui/material/TextField";
 import InputAdornment from "@mui/material/InputAdornment";
 import SearchIcon from "@mui/icons-material/Search";
-import { TopBar } from "../../components/TopBar";
+import { TopBar } from "../../../components/TopBar";
 import Box from "@mui/material/Box";
-import { Typography, Button, Portal, FormGroup } from "@mui/material";
-import { LoadingButton } from "@mui/lab";
-import {
-  useCreateHeadTagMutation,
-  useUpdateHeadTagsMutation,
-} from "../../../../../../shell/services/instance";
-import AddIcon from "@mui/icons-material/Add";
-import { notify } from "../../../../../../shell/store/notifications";
-import { FormControlLabel, Checkbox } from "@mui/material";
-import { useDispatch, useSelector } from "react-redux";
+import { Typography, Button, Portal } from "@mui/material";
+import { NoResults } from "../../../../../../schema/src/app/components/NoResults";
+import SearchBox from "../../../../../../../shell/components/SearchBox";
+import FontFamilyCard from "./FontFamilyCard";
+import { useSettingsFonts } from "../hooks/useSettingsFonts";
 
-import { NoResults } from "../../../../../schema/src/app/components/NoResults";
-import { AppState } from "../../../../../../shell/store/types";
-import SearchBox from "../../../../../../shell/components/SearchBox";
-import { useSettingsFonts } from "../../components/useSettingsFonts";
-
-export type FontRowItemProps = {
-  ZUID: string | null;
-  family: string;
-  variants: string[];
-  installedVariants: string[];
-  previewText: string;
-  activePage: number;
-};
 type InstalledFont = {
   ZUID: string;
   family: string;
   variants: string[];
   href: string;
 };
-const FontRowItem: FC<FontRowItemProps> = ({
-  ZUID = null,
-  family,
-  variants,
-  installedVariants,
-  previewText,
-  activePage,
-}) => {
-  const dispatch = useDispatch();
-  const instance = useSelector((state: AppState) => state.instance);
 
-  const [selectedVariants, setSelectedVariants] = useState([]);
-  const [activeVariants, setActiveVariants] = useState(installedVariants);
-
-  const {
-    installedFonts,
-    isLoading: isLoadingInstalledFonts,
-    getFontDataFromHref,
-  } = useSettingsFonts();
-
-  const [createHeadTag, { isLoading: isCreating }] = useCreateHeadTagMutation();
-  const [updateFont, { isLoading: isUpdating }] = useUpdateHeadTagsMutation();
-
-  const isLoading = isCreating || isUpdating;
-
-  const handleSelectionChange = (value: boolean, item: string) => {
-    if (value) {
-      setSelectedVariants((prev) => [...prev, item]);
-    } else {
-      setSelectedVariants((prev) => prev.filter((variant) => variant !== item));
-    }
-  };
-
-  const handleFontInstall = useCallback(async () => {
-    try {
-      const cssLinkUrl = "https://fonts.googleapis.com/css?family=";
-
-      const currentlyInstalledVariants = !installedVariants
-        ? selectedVariants
-        : [...installedVariants, ...selectedVariants];
-
-      const newFontName = family?.trim()?.replace(/\s/g, "+");
-      const linkHref = `${cssLinkUrl}${newFontName}:${currentlyInstalledVariants?.join(
-        ","
-      )}`;
-
-      let response: any = null;
-      if (!ZUID) {
-        response = await createHeadTag({
-          type: "link",
-          resourceZUID: instance.ZUID,
-          sort: 0,
-          attributes: {
-            rel: "stylesheet",
-            href: linkHref,
-          },
-        });
-      } else {
-        response = await updateFont({
-          ZUID,
-          href: linkHref,
-        });
-      }
-
-      // dispatch(fetchFontsInstalled());
-
-      if (!response?.error) {
-        setSelectedVariants([]);
-        setActiveVariants(currentlyInstalledVariants);
-        dispatch(
-          notify({
-            kind: "success",
-            message: `Font "${family} (${selectedVariants.join(
-              ", "
-            )})" has been installed`,
-          })
-        );
-      } else {
-        throw new Error(`${response?.error?.data?.error}`);
-      }
-    } catch (error) {
-      dispatch(
-        notify({
-          kind: "error",
-          message: `Failed to add ${family} (${selectedVariants.join(
-            ", "
-          )}): ${error}`,
-        })
-      );
-    }
-  }, [ZUID, selectedVariants, installedVariants]);
-
-  useEffect(() => {
-    setSelectedVariants([]);
-  }, [activePage]);
-
-  return (
-    <>
-      <Box
-        sx={{
-          borderBottom: "1px solid",
-          borderColor: "grey.200",
-          pb: 2,
-          pt: 3,
-        }}
-      >
-        <Box
-          sx={{
-            display: "flex",
-            justifyContent: "space-between",
-            alignItems: "flex-start",
-            columnGap: 3,
-          }}
-        >
-          <Box
-            sx={{
-              display: "flex",
-              justifyContent: "space-between",
-              alignItems: "flex-start",
-              flexDirection: "column",
-              width: "100%",
-              rowGap: 1.25,
-            }}
-          >
-            <Typography variant="h5" fontWeight={700} color="text.primary">
-              {family}
-            </Typography>
-            <Box
-              sx={{
-                lineHeight: 2,
-                textTransform: "capitalize",
-                display: "flex",
-                flexWrap: "wrap",
-              }}
-            >
-              <FormGroup row>
-                {variants?.map((item: string, index: number) => (
-                  <FormControlLabel
-                    key={`${family}-${item}-${ZUID || index}`}
-                    color="text.primary"
-                    control={
-                      <Checkbox
-                        key={`${family}-${item}-${ZUID || index}`}
-                        color="primary"
-                        size="small"
-                        id={`${family}-${item}-${ZUID || index}`}
-                        checked={
-                          (selectedVariants?.includes(item) ||
-                            activeVariants?.includes(item)) ??
-                          false
-                        }
-                        disabled={activeVariants?.includes(item)}
-                        readOnly={activeVariants?.includes(item)}
-                        onChange={(_e, val) => handleSelectionChange(val, item)}
-                        sx={{ p: 0 }}
-                      />
-                    }
-                    label={item}
-                    labelPlacement="end"
-                    slotProps={{
-                      typography: {
-                        variant: "body2",
-                        fontWeight: 400,
-                        color: "text.secondary",
-
-                        pr: 0.25,
-                        pl: 0,
-                        width: "60px",
-                      },
-                    }}
-                    sx={{
-                      m: 0,
-                    }}
-                  />
-                ))}
-              </FormGroup>
-            </Box>
-          </Box>
-
-          <LoadingButton
-            variant="contained"
-            color="primary"
-            onClick={handleFontInstall}
-            disabled={!isLoading && !selectedVariants?.length}
-            startIcon={<AddIcon />}
-            loading={isLoading}
-            loadingPosition="center"
-            sx={{
-              minWidth: "fit-content",
-            }}
-          >
-            Add
-          </LoadingButton>
-        </Box>
-        <Typography
-          variant="h4"
-          fontWeight={400}
-          color="text.primary"
-          sx={{ fontFamily: `"${family}"`, mt: 1 }}
-        >
-          {previewText
-            ? previewText
-            : "All their equipment and instruments are alive."}
-        </Typography>
-      </Box>
-    </>
-  );
-};
-const Browser = () => {
+const Browse = () => {
   const [loadComplete, setLoadComplete] = useState<boolean>(false);
   const [search, setSearch] = useState<string>("");
   const searchInputRef = useRef<HTMLInputElement>(null);
@@ -262,11 +36,11 @@ const Browser = () => {
       return [];
     } else {
       const paginationWithZUID = [...webFonts]
-        ?.sort((a: any, b: any) => a?.family?.localeCompare(b?.family))
+        ?.sort((a, b) => a?.family?.localeCompare(b?.family))
         ?.map((gFont, index) => {
           const currentInstalledFonts = installedFonts?.find(
             (item: InstalledFont) => {
-              const { family, variants } = getFontDataFromHref(item?.href);
+              const { family } = getFontDataFromHref(item?.href);
               return family === gFont?.family;
             }
           );
@@ -336,16 +110,6 @@ const Browser = () => {
           alignItems="center"
           justifyContent="space-between"
           columnGap={2}
-          sx={{
-            "& .MuiInputBase-root fieldset": {
-              bgcolor: "grey.50",
-            },
-            "& .MuiInputBase-root": {
-              "& input, & .MuiInputAdornment-root": {
-                zIndex: 5,
-              },
-            },
-          }}
         >
           <SearchBox
             placeholder="Search Fonts"
@@ -370,22 +134,18 @@ const Browser = () => {
               setSearch(evt.target.value);
             }}
             inputRef={searchInputRef}
-            slotProps={{
-              input: {
-                startAdornment: (
-                  <InputAdornment position="start">
-                    <SearchIcon fontSize="small" />
-                  </InputAdornment>
-                ),
+            InputProps={{
+              startAdornment: (
+                <InputAdornment position="start">
+                  <SearchIcon fontSize="small" />
+                </InputAdornment>
+              ),
+              sx: {
+                bgcolor: "grey.50",
               },
             }}
             sx={{
               width: "280px",
-              "& .MuiInputBase-root.MuiOutlinedInput-root.MuiInputBase-sizeSmall":
-                {
-                  py: 0.5,
-                  bgcolor: "grey.50",
-                },
             }}
           />
         </Box>
@@ -399,7 +159,7 @@ const Browser = () => {
           height: "calc(100% - 84px)",
           overflowY: "auto",
           overflowX: "hidden",
-          margin: "0",
+          margin: 0,
           display: "block",
           maxHeight: "calc(100% - 84px)",
           position: "relative",
@@ -465,7 +225,7 @@ const Browser = () => {
           ) : (
             <>
               {activePagination?.map((itemFont, index) => (
-                <FontRowItem
+                <FontFamilyCard
                   previewText={previewText}
                   key={itemFont?.family + index}
                   family={itemFont?.family}
@@ -536,4 +296,4 @@ const Browser = () => {
     </>
   );
 };
-export default Browser;
+export default Browse;
